(refactor) Better name
